Remove https:// from custom url pattern regex
The regex pattern for finding custom urls like 'call.disqus.com' had 'https://' in it so the event description strings were not yielding any valid matches, which resulted in hangoutlooks creating a new call URL instead of using the provided. Also adds 'http://' to the callURL when the custom url pattern is found so that the meeting event href works.